% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accMass.R
\name{accMass}
\alias{accMass}
\title{Collect accurate masses(5dp) for a given list of explanatory masses(2dp) based on the mosted intense signal}
\usage{
accMass(master_mat, cls, masses)
}
\description{
Collect accurate masses(5dp) for a given list of explanatory masses(2dp) based on the mosted intense signal
}

