% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accMassFinal.R
\name{accMassFinal}
\alias{accMassFinal}
\title{Make final accurate mass matrix from acc_mat using the most intense mass}
\usage{
accMassFinal(acc_mat)
}
\description{
Make final accurate mass matrix from acc_mat using the most intense mass
}

