% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxLine.R
\name{boxLine}
\alias{boxLine}
\title{boxplot and line plotting function, masses=NULL plots all}
\usage{
boxLine(dat.all, cls, Path, DF, dn, masses, type = "bl", h.group = NULL,
  v.group = NULL, HPC_mode = F)
}
\description{
boxplot and line plotting function, masses=NULL plots all
}

