% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corFeatures.R
\name{corFeatures}
\alias{corFeatures}
\title{correlation dendrogram of explanatory features based on all the data.}
\usage{
corFeatures(dat.all, masses)
}
\description{
correlation dendrogram of explanatory features based on all the data.
}

