% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmapRatios.R
\name{heatmapRatios}
\alias{heatmapRatios}
\title{Plot heatmap of log2 ratios treatments}
\usage{
heatmapRatios(data, masses, cls.treat, cls.time, rel = T)
}
\description{
Plot heatmap of log2 ratios treatments
}

