% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ldaAnalysis.R
\name{ldaAnalysis}
\alias{ldaAnalysis}
\title{wrapper function for LDA plotting, returns a list containg results.}
\usage{
ldaAnalysis(dat.all, cls, Path, DF, cls.1 = NULL, HPC_mode = F)
}
\description{
wrapper function for LDA plotting, returns a list containg results.
}

