% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldAnalysis.R
\name{ldAnalysis}
\alias{ldAnalysis}
\title{wrapper function for LDA plotting, returns a list containg results.}
\usage{
ldAnalysis(dat.all, cls)
}
\description{
wrapper function for LDA plotting, returns a list containg results.
}

