% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linePlot.R
\name{linePlot}
\alias{linePlot}
\title{boxplot and line plotting function, masses=NULL plots all}
\usage{
linePlot(dat, h.group, v.group, mass)
}
\description{
boxplot and line plotting function, masses=NULL plots all
}

