% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listsPrepSig.R
\name{listsPrepSig}
\alias{listsPrepSig}
\title{return lists of explanatory masses for each pairwise}
\usage{
listsPrepSig(fs.res, meth, cut, sel = NULL)
}
\description{
return lists of explanatory masses for each pairwise
}

