% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterMat.R
\name{masterMat}
\alias{masterMat}
\title{to build mean intensity matrices  by class including nominal and 2dp bins}
\usage{
masterMat(master_mix, cls)
}
\description{
to build mean intensity matrices  by class including nominal and 2dp bins
}

