% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/occDrop.R
\name{occDrop}
\alias{occDrop}
\title{drop variables with occupancy below percentage threshold}
\usage{
occDrop(peak.mat, cls, perc)
}
\description{
drop variables with occupancy below percentage threshold
}

