% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outDet.R
\name{outDet}
\alias{outDet}
\title{changed so that actual sample labels used}
\usage{
outDet(x, samp.names, method = "classical", conf.level = 0.975,
  dimen = c(1, 2), tol = 1e-07, plotting = TRUE)
}
\description{
changed so that actual sample labels used
}

