% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcaAnalysis.R
\name{pcaAnalysis}
\alias{pcaAnalysis}
\title{wrapper function for PCA plotting, returns a list containg results}
\usage{
pcaAnalysis(dat.all, cls, Path, DF, cls.1 = NULL, HPC_mode = F)
}
\description{
wrapper function for PCA plotting, returns a list containg results
}

