% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcAnalysis.R
\name{pcAnalysis}
\alias{pcAnalysis}
\title{wrapper function for principle component analysis, returns a list containg results}
\usage{
pcAnalysis(dat.all, cls)
}
\description{
wrapper function for principle component analysis, returns a list containg results
}

