% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBin.R
\name{plotBin}
\alias{plotBin}
\title{Plot peak intensities for a 2dp mass bins for all classes}
\usage{
plotBin(master_mat, masses, Path, DF)
}
\description{
Plot peak intensities for a 2dp mass bins for all classes
}

