% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProc.R
\name{preProc}
\alias{preProc}
\title{FIE-HRMS data pre-processing}
\usage{
preProc(dat, cls, proportion, add = 1)
}
\arguments{
\item{dat}{A data.frame}

\item{cls}{A vector denoting the class of each observation in the data.frame dat.}

\item{proportion}{The proportion threshold between 0 and 1, above which a varible must contain at least one class.}

\item{add}{A numeric value for addition to the matrix prior to log10 transformation.}
}
\value{
A pre-processed data.frame
}
\description{
Pre-process FIE-HRMS data suitable for downstream analysis.
}
\details{
Includes TIC normalisation, class occupancy filtering and log10 transformation in that order.
}
\author{
Jasen Finch
}

