% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDat.R
\name{processDat}
\alias{processDat}
\title{processDat}
\usage{
processDat(peak.mat, cls = NULL, bat = NULL, out_id = NULL,
  out_idx = NULL, nCores = 2, pre_SD = T, mean_shift = T)
}
\description{
processDat
}

