% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfImportance.R
\name{rfImportance}
\alias{rfImportance}
\title{function to plot RF importance curves}
\usage{
rfImportance(fs.res, sel, Path, DF, meth = "rf", HPC_mode = F)
}
\description{
function to plot RF importance curves
}

