% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionFrequencyFPR.R
\name{selectionFrequencyFPR}
\alias{selectionFrequencyFPR}
\title{false positive approximation for random forest selection frequencies}
\usage{
selectionFrequencyFPR(k, K, Tr, Ft)
}
\arguments{
\item{k}{variable selction frequency}

\item{K}{The average number of binary tests made across a forest}

\item{Tr}{The number of trees used to build the forest}

\item{Ft}{Total number of variables}
}
\value{
A numeric value giving the false positive rate
}
\description{
false positive approximation for random forest selection frequencies
}
\author{
Jasen Finch
}

