% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryDat.R
\name{summaryDat}
\alias{summaryDat}
\title{Function for calculating confidence intervals and errors for plots}
\usage{
summaryDat(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\description{
Function for calculating confidence intervals and errors for plots
}

