% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewSpectrum.R
\name{viewSpectrum}
\alias{viewSpectrum}
\title{view zoom able spectrum. double click to zoom out. 2 single clicks over range to zoom in.}
\usage{
viewSpectrum(mat, class, mode, type = "l")
}
\description{
view zoom able spectrum. double click to zoom out. 2 single clicks over range to zoom in.
}

