% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listsPrepSig.R
\name{listsPrepSig}
\alias{listsPrepSig}
\title{return lists of explanatory masses for each pairwise}
\usage{
listsPrepSig(fs.res, DF, meth, sel, cut, dn)
}
\description{
return lists of explanatory masses for each pairwise
}

