% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occDrop.R
\name{occDrop}
\alias{occDrop}
\title{Class occupancy filtering.}
\usage{
occDrop(dat, cls, proportion)
}
\arguments{
\item{dat}{A data.frame.}

\item{cls}{A vector denoting the class of each observation in the data.frame dat.}

\item{proportion}{The proportion threshold between 0 and 1, above which a varible must contain at least one class.}
}
\value{
An occupancy filtered data.frame
}
\description{
Drop variables without a maximum class occupancy above a given proportion.
}
\author{
Jasen Finch
}

