% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rfImportance.R
\name{rfImportance}
\alias{rfImportance}
\title{function to plot RF importance curves}
\usage{
rfImportance(fs.res, dn, sel, Path, DF, meth = "rf")
}
\description{
function to plot RF importance curves
}

