% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TICnorm.R, R/fs.rf2.R
\name{TICnorm}
\alias{TICnorm}
\alias{fs.rf2}
\title{Total Ion Count Normalisation.}
\usage{
TICnorm(dat)

fs.rf2(train, cls, ...)
}
\arguments{
\item{dat}{A data.frame.}

\item{train}{training set data}

\item{cls}{training set class labels}
}
\value{
A TIC normalised data.frame.

A list containing selection frequency feature selection results
}
\description{
Normalise FIE-HRMS spectra to their total ion count.

Normalise FIE-HRMS spectra to their total ion count.
}
\author{
Jasen Finch

Jasen Finch
}

