% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TICnorm.R
\name{TICnorm}
\alias{TICnorm}
\title{Total Ion Count Normalisation.}
\usage{
TICnorm(dat)
}
\arguments{
\item{dat}{A data.frame.}
}
\value{
A TIC normalised data.frame.
}
\description{
Normalise FIE-HRMS spectra to their total ion count.
}
\author{
Jasen Finch
}

